requirements.txt
astropy==4.0.1.post1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
click==7.1.1
colorama==0.4.1
confuse==1.1.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.2
django-cors-headers==3.2.1
django-filter==2.2.0
django-filters==0.2.1
django-rest-framework==0.1.0
djangorestframework==3.10.2
dnspython==1.16.0
entrypoints==0.3
fastapi==0.52.0
gunicorn==20.0.4
h11==0.9.0
html5lib==1.0.1
htmlmin==0.1.12
idna==2.8
image==1.5.27
ImageHash==4.1.0
inflection==0.4.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.11.2
joblib==0.13.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
llvmlite==0.32.1
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
missingno==0.4.2
mistune==0.8.4
more-itertools==8.2.0
mpl-finance==0.10.1
mplfinance==0.12.4a0
mysql-connector-python==8.0.16
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.4
notebook==6.0.1
numba==0.49.1
numpy==1.17.2
pandas==0.25.3
pandas-datareader==0.8.1
pandas-profiling==2.8.0
pandocfilters==1.4.2
parso==0.5.1
phik==0.10.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
pydantic==1.4
Pygments==2.4.2
PyJWT==1.7.1
pymongo==3.10.1
PyMySQL==0.9.3
pyparsing==2.4.2
pyrsistent==0.15.4
pyserial==3.4
pytesseract==0.3.4
python-dateutil==2.8.0
pyttsx3==2.71
pytz==2019.1
PyWavelets==1.1.1
pywinpty==0.5.5
PyYAML==5.3.1
pyzmq==18.1.0
qtconsole==4.5.5
Quandl==3.5.0
requests==2.23.0
scikit-learn==0.21.3
scipy==1.4.1
seaborn==0.10.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
soupsieve==1.9.3
SpeechRecognition==3.8.1
sqlparse==0.3.0
starlette==0.13.2
tangled-up-in-unicode==0.0.6
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.46.1
traitlets==4.3.2
urllib3==1.25.3
uvicorn==0.11.3
visions==0.4.4
wcwidth==0.1.7
webencodings==0.5.1
websockets==8.1
widgetsnbextension==3.5.1
wikipedia==1.4.0
xlrd==1.2.0
